Sketcher: Excise SketchGeometryExtension outside of SketchObject
