+++ src/Mod/Sketcher/App/SketchObject.cpp
@@ -89,59 +89,6 @@
 using namespace Sketcher;
 using namespace Base;
 
-//---------- Geometry Extension
-
-TYPESYSTEM_SOURCE(Sketcher::SketchGeometryExtension,Part::GeometryExtension)
-
-SketchGeometryExtension::SketchGeometryExtension():id(0)
-{
-
-}
-
-SketchGeometryExtension::SketchGeometryExtension(long cid):id(cid)
-{
-
-}
-
-SketchGeometryExtension::~SketchGeometryExtension()
-{
-}
-
-// Persistence implementer
-unsigned int SketchGeometryExtension::getMemSize (void) const
-{
-    return sizeof(long int);
-}
-
-void SketchGeometryExtension::Save(Base::Writer &writer) const
-{
-
-    writer.Stream() << writer.ind() << "<GeoExtension type=\"" << this->getTypeId().getName()
-                    << "\" id=\"" << id << "\"/>" << endl;
-}
-
-void SketchGeometryExtension::Restore(Base::XMLReader &reader)
-{
-    id = reader.getAttributeAsInteger("id");
-}
-
-std::unique_ptr<Part::GeometryExtension> SketchGeometryExtension::copy(void) const
-{
-    std::unique_ptr<SketchGeometryExtension> cpy = std::make_unique<SketchGeometryExtension>();
-
-    cpy->id = this->id;
-
-    return std::move(cpy);
-}
-
-PyObject * SketchGeometryExtension::getPyObject(void)
-{
-    return new SketchGeometryExtensionPy(new SketchGeometryExtension(this->id));
-}
-
-
-//-------------- End Geometry Extension
-
 
 const int GeoEnum::RtPnt  = -1;
 const int GeoEnum::HAxis  = -1;
